Refactor compose theme package
